requirements.txt
alembic==1.6.5
amqp==5.0.6
anyio==3.3.0
argon2-cffi==20.1.0
asgi-lifespan==1.0.1
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
boto3==1.18.26
botocore==1.21.26
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==5.5
cryptography==3.4.7
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
Faker==8.12.0
fastapi==0.68.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.6
httpx==0.19.0
idna==3.2
iniconfig==1.1.1
jmespath==0.10.0
kombu==5.1.0
Mako==1.1.4
MarkupSafe==2.0.1
packaging==21.0
passlib==1.7.4
pluggy==0.13.1
prompt-toolkit==3.0.20
psycopg2==2.9.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.4
pytest-asyncio==0.15.1
pytest-mock==3.6.1
python-dateutil==2.8.2
python-decouple==3.6
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
redis==3.5.3
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.24
starlette==0.14.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
vine==5.0.0
wcwidth==0.2.5